Skipped running finished calculations

+++ Code/Tools/MD/Umbrella.py
@@ -273,13 +273,14 @@ nohup qmmm_sniffer & diswon
 """
             
             for bin in self.data.keys():
-                if HPC.exists:
-                    if MM.qm == "pyscf":
-                        runscript += f"cd {bin} ; sed -i \"s/RUNDIR/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/g\" {job['output']}.conf ; mkdir /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID; mkdir /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/0 ; cd /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/0 ; nohup qmmm_sniffer & disown ; cd - ; {MMPath} {CommandLines} {job['output']}.conf > {job['output']}.out ; cd ../ ; touch /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/0/kill ; sleep 10 ; rm -r /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID ;\n"
+                if MM.software.check_output(f"{bin}/{job['output']}.out")[0] != "completed":
+                    if HPC.exists:
+                        if MM.qm == "pyscf":
+                            runscript += f"cd {bin} ; sed -i \"s/RUNDIR/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/g\" {job['output']}.conf ; mkdir /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID; mkdir /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/0 ; cd /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/0 ; nohup qmmm_sniffer & disown ; cd - ; {MMPath} {CommandLines} {job['output']}.conf > {job['output']}.out ; cd ../ ; touch /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/0/kill ; sleep 10 ; rm -r /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID ;\n"
+                        else:
+                            runscript += f"cd {bin} ; sed -i \"s/RUNDIR/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/g\" {job['output']}.conf ; mkdir /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID ; {MMPath} {CommandLines} {job['output']}.conf > {job['output']}.out ; cd ../ ; rm -r /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID ;\n"
                     else:
-                        runscript += f"cd {bin} ; sed -i \"s/RUNDIR/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/g\" {job['output']}.conf ; mkdir /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID ; {MMPath} {CommandLines} {job['output']}.conf > {job['output']}.out ; cd ../ ; rm -r /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID ;\n"
-                else:
-                    runscript += f"cd {bin} ; {MMPath} {CommandLines} {job['output']}.conf > {job['output']}.out ; cd ../ \n"
+                        runscript += f"cd {bin} ; {MMPath} {CommandLines} {job['output']}.conf > {job['output']}.out ; cd ../ \n"
             if HPC.exists == False:
                 if MM.qm == "pyscf":
                     runscript += "touch /dev/shm/RUNDIR/0/kill"
@@ -311,10 +312,11 @@ nohup qmmm_sniffer & diswon
         """
                 
                 for bin in self.data.keys():
-                    if HPC.exists:
-                        runscript += f"cd {bin} ; sed -i \"s/RUNDIR/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/g\" {job['output']}_{i+1}.conf ; mkdir /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID ; {MMPath} {CommandLines} {job['output']}_{i+1}.conf > {job['output']}_{i+1}.out ; cd ../ ; rm -r /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID ;\n"
-                    else:
-                        runscript += f"cd {bin} ; {MMPath} {CommandLines} {job['output']}_{i+1}.conf > {job['output']}_{i+1}.out ; cd ../ \n"
+                    if MM.software.check_output(f"{bin}/{job['output']}_{i+1}.out")[0] != "completed":
+                        if HPC.exists:
+                            runscript += f"cd {bin} ; sed -i \"s/RUNDIR/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID/g\" {job['output']}_{i+1}.conf ; mkdir /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID ; {MMPath} {CommandLines} {job['output']}_{i+1}.conf > {job['output']}_{i+1}.out ; cd ../ ; rm -r /dev/shm/$SLURM_JOB_ID-$SLURM_ARRAY_TASK_ID ;\n"
+                        else:
+                            runscript += f"cd {bin} ; {MMPath} {CommandLines} {job['output']}_{i+1}.conf > {job['output']}_{i+1}.out ; cd ../ \n"
                 if HPC.exists == False:
                     runscript += "rm -r /dev/shm/RUNDIR"
